requirements.txt
altgraph==0.17
amqp==2.6.0
arrow==0.14.5
asgiref==3.2.7
atomicwrites==1.3.0
attrs==19.3.0
Babel==2.8.0
billiard==3.6.3.0
boto==2.49.0
boto3==1.13.20
botocore==1.16.20
celery==4.3.0
certifi==2020.4.5.1
cffi==1.14.2
chardet==3.0.4
codacy-coverage==1.3.6
colorama==0.4.3
compress==0.0.2
coverage==5.1
cryptography==3.1
cssselect==1.1.0
defusedxml==0.6.0
dj-database-url==0.4.2
Django==3.0.1
django-appconf==1.0.4
django-ckeditor==5.9.0
django-cleanup==4.0.0
django-compressor==2.4
django-crispy-forms==1.8.0
django-crm==0.8.0
django-filebrowser==3.9.1
django-filter==1.1.0
django-grappelli==2.10.1
django-js-asset==1.2.2
django-model-utils==4.0.0
django-notification==1.3.3
django-notifications-hq==1.6.0
django-phonenumber-field==4.0.0
django-ranged-response==0.2.0
django-settings-export==1.2.1
django-simple-captcha==0.5.12
django-simple-pagination==1.3
django-storages==1.9.1
django-user-accounts==3.0.2
djangorestframework==3.11.0
docutils==0.15.2
elasticsearch==7.7.1
et-xmlfile==1.0.1
factory-boy==2.11.1
Faker==4.1.2
future==0.18.2
get==2019.4.13
gunicorn==19.7.1
idna==2.8
importlib-metadata==1.6.0
jdcal==1.4.1
jmespath==0.10.0
jsonfield==3.1.0
kombu==4.6.9
libsass==0.20.0
lxml==4.4.1
Markdown==2.6.11
more-itertools==8.5.0
oauthlib==3.1.0
olefile==0.44
openpyxl==3.0.3
packaging==20.3
pdfkit==0.6.1
pefile==2019.4.18
phonenumbers==8.12.9
Pillow==7.1.2
pluggy==0.13.1
post==2019.4.13
psycopg2==2.8.4
public==2019.4.13
py==1.8.1
pycparser==2.20
PyInstaller==3.6
PyJWT==1.7.1
pyparsing==2.4.7
pytest==3.7.0
pytest-cov==2.5.1
pytest-django==3.1.2
python-dateutil==2.8.1
python-dotenv==0.14.0
python-memcached==1.59
python3-openid==3.1.0
pytz==2019.3
pywin32-ctypes==0.2.0
query-string==2019.4.13
raven==6.10.0
redis==3.3.8
reportlab==3.5.34
request==2019.4.13
requests==2.22.0
requests-oauthlib==1.3.0
rjsmin==1.1.0
s3transfer==0.3.3
selenium==3.14.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.2.0
sorl-thumbnail==12.6.3
sqlparse==0.3.1
swapper==1.1.2.post1
text-unidecode==1.3
urllib3==1.25.9
vine==1.3.0
wcwidth==0.1.9
whitenoise==5.1.0
xlrd==1.2.0
xlwt==1.3.0
zipp==3.1.0
